+++ DependencyInjection/Factory/Resolver/AwsS3ResolverFactory.php
@@ -103,7 +103,6 @@ class AwsS3ResolverFactory extends AbstractResolverFactory
                 ->end()
                 ->scalarNode('acl')
                     ->defaultValue('public-read')
-                    ->cannotBeEmpty()
                 ->end()
                 ->scalarNode('cache_prefix')
                     ->defaultValue('')