+++ CHANGELOG.md
@@ -9,6 +9,7 @@ for a given releases. Unreleased, upcoming changes will be updated here periodic
 ## Next Release
 
 - Add `use_psr_cache` option to AWS S3 resolver configuration in order to support PSR cache. Configuring a doctrine cache now will trigger a deprecation ([deguif](https://github.com/liip/LiipImagineBundle/pull/1630))
+- Add ability to specify a custom s3 client service in aws s3 resolver configuration ([deguif](https://github.com/liip/LiipImagineBundle/pull/1632))
 
 ## [2.13.3](https://github.com/liip/LiipImagineBundle/tree/2.13.3)